plug-ins: Quiet some 64-bit to 32-bit warnings on 64-bit OSes, specifically OS X.
git-svn-id: https://pcsxr.svn.codeplex.com/svn/pcsxr@91796 e17a0e51-4ae3-4d35-97c3-1a29b211df97

+++ plugins/bladesio1/connection.c
@@ -138,7 +138,7 @@ int connectionSend(u8 *pdata, s32 size) {
 	int bytes = 0;
 
 	if(clientsock >= 0)
-		if((bytes = send(clientsock, (const char*)pdata, size, 0)) < 0)
+		if((bytes = (int)send(clientsock, (const char*)pdata, size, 0)) < 0)
 			return 0;
 
 	return bytes;
@@ -148,7 +148,7 @@ int connectionRecv(u8 *pdata, s32 size) {
 	int bytes = 0;
 
 	if(clientsock >= 0)
-		if((bytes = recv(clientsock, (char*)pdata, size, 0)) < 0)
+		if((bytes = (int)recv(clientsock, (char*)pdata, size, 0)) < 0)
 			return 0;
 
 	return bytes;
